refactor(harness): fix session resume bugs and extract hiclaw-common shared package
## PR #1 — Bug fixes

- fix(harness): matrix_relay.py always passed None as session_id to on_invoke,
  so claude --resume was never triggered. Narrowed callback type to Callable[[str], ...]
  and removed the hardcoded None at the call site.
- fix(harness): _load_session() was defined but never called. Worker now loads
  the persisted session id at startup and threads it through a closure so
  every harness invocation resumes the prior Claude conversation after restart.

## PR #2 — hiclaw-common shared package + mautrix migration

- feat(shared): introduce HiClaw/shared/python/hiclaw_common package with three
  modules extracted from harness/hermes verbatim-copy duplication:
    - policies.py  — DualAllowList, HistoryBuffer, apply_outbound_mentions (MSC3952)
    - sync.py      — FileSync, sync_loop, push_loop, _deep_merge, _merge_openclaw_config;
                     parameterized via runtime_home_dir (.harness/.hermes) so both
                     runtimes share the same implementation
    - matrix.py    — new mautrix-based MautrixRelay + matrix_relogin helper; converges
                     harness onto the same Matrix SDK (mautrix) hermes already uses
- refactor(harness): replace 135-line matrix-nio MatrixRelay with a 40-line thin adapter
  over MautrixRelay from hiclaw_common
- refactor(harness): sync.py is now a thin re-export from hiclaw_common.sync
- delete(harness): harness_worker/policies.py (verbatim copy removed)
- chore(harness): swap nio>=0.24 dep for hiclaw-common[matrix] in pyproject.toml
- chore(harness/Dockerfile): install hiclaw_common[matrix] from shared-python
  named build context (--build-context shared-python=./shared/python)
- chore(Makefile): add SHARED_PYTHON_CTX; wire into harness build + push targets

## PR #3 — Hermes cutover to hiclaw_common.policies

- refactor(hermes): overlay_adapter.py imports DualAllowList/HistoryBuffer/
  apply_outbound_mentions from hiclaw_common.policies instead of hermes_matrix.policies
- delete(hermes): hermes_matrix/policies.py (canonical copy now in hiclaw_common)
- chore(hermes): add hiclaw-common>=0.1.0 to pyproject.toml
- chore(hermes/Dockerfile): install hiclaw_common from shared-python build context
  before hermes-agent; wire $(SHARED_PYTHON_CTX) into hermes build + push targets
