+++ harness/pyproject.toml
@@ -11,7 +11,7 @@ dependencies = [
     "typer>=0.12",
     "rich>=13.0",
     "httpx>=0.27",
-    "nio>=0.24",
+    "hiclaw-common[matrix] @ file:///../shared/python/hiclaw_common",
 ]
 
 [project.scripts]
