+++ shared/python/hiclaw_common/src/hiclaw_common/__init__.py
@@ -1,0 +1,3 @@
+"""HiClaw shared runtime capabilities."""
+
+__version__ = "0.1.0"
